refactor: 퀴즈 선택지 컴포넌트 추상화
