+++ src/app/(routes)/quiz/components/multiple-options/components/option.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { cn } from '@/lib/utils'
 import { VariantProps, cva } from 'class-variance-authority'
