+++ src/app/(routes)/quiz/components/quiz-header.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { Button } from '@/components/ui/button'
 import { useRouter } from 'next/navigation'
 import { HTMLAttributes } from 'react'
