+++ src/app/(routes)/quiz/types/index.tsx
@@ -1,0 +1,6 @@
+export interface QuizProgress {
+  quizIndex: number
+  selectedMultipleQuizAnswer: number | null
+  selectedMixUpQuizAnswer: 'correct' | 'incorrect' | null
+  progress: 'idle' | 'choose' | 'result'
+}
